fix broken once.remove
