+++ package.json
@@ -7,12 +7,12 @@
     "type": "git",
     "url": "git://github.com/ssbc/obz.git"
   },
-  "dependencies": {},
   "devDependencies": {
-    "tape": "^4.6.0"
+    "tap-arc": "^1.0.1",
+    "tape": "^5.7.0"
   },
   "scripts": {
-    "test": "set -e; for t in test/*.js; do node $t; done"
+    "test": "tape test/*.js | tap-arc"
   },
   "author": "'Dominic Tarr' <dominic.tarr@gmail.com> (dominictarr.com)",
   "license": "MIT"
