+++ test/trigger.js
@@ -167,56 +167,6 @@ module.exports = function (observable) {
     t.end()
   })
 
-  tape('once', function (t) {
-    var o = observable()
-    var fired = 0
-    o.once(function (v) {
-      fired ++
-      t.equal(v, 1)
-    })
-
-    o.set(1)
-
-    o.once(function (v) {
-      fired ++
-      t.equal(v, 2)
-    }, false)
-
-    o.once(function (v) {
-      fired ++
-      t.equal(v, 1)
-    })
-
-    t.equal(fired, 2)
-
-    o.set(2)
-
-    o.set(3)
-
-    t.equal(fired, 3)
-
-    t.end()
-  })
-
-  tape('once, !immediately, inside trigger', function (t) {
-    var o = observable()
-    var fired = 0
-    o.once(function (v) {
-      t.equal(v, 1)
-      fired ++
-      o.once(function (v) {
-        fired ++
-        t.equal(v, 2)
-      }, false)
-    })
-
-    o.set(1)
-    t.equal(fired, 1)
-    o.set(2)
-    t.equal(fired, 2)
-    t.end()
-  })
-
   tape('bug', (t) => {
     t.plan(4)
     var o = observable()
@@ -251,5 +201,8 @@ module.exports = function (observable) {
 
 }
 
-if(!module.parent) module.exports (require('../'))
+if(
+  !module.parent ||
+  module.parent.path.endsWith('node_modules/tape/bin')
+) module.exports(require('../'))
